Added implicit map parsing mode
This commit adds the concept of an implicit map -- detecting where the
root starts with `Identifier Colon` and switching into map parsing.

I commented on #6 about the remaining "edge case" I hadn't thought of --
handling of commas vs new lines in this mode. Currently, commas are
still required, and whitespace is insignificant, but this likely will be
updated to support removing commas between key-value pairs as long as
there's an end-of-line.