Refactored newtype support
Honestly, I don't even want to try to summarize what I did. To try to
make all these cases work, it took careful management of state to make
work. I'm not sure if this will break with further nestings or not, but
it covers the positive flow cases in the unit test which I think are the
real-world practical use cases for this.